pyproject.toml
[tool.poetry]
name = "drf-hw"
version = "0.1.0"
description = ""
authors = ["Ramzes_13 <rslobolinsky13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
asgiref = "^3.8.1"
django-filter = "^24.2"
djangorestframework = "^3.15.1"
pillow = "^10.3.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
sqlparse = "^0.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
